LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/constructor-expected.txt


FAIL Animation can be constructed with null effect and null timeline assert_equals: Animation timeline should be null expected null but got object "[object DocumentTimeline]"
PASS Animation can be constructed with null effect and non-null timeline 
PASS Animation can be constructed with null effect and no timeline parameter 
FAIL Animation can be constructed with non-null effect and null timeline Can't find variable: KeyframeEffectReadOnly
FAIL Animation can be constructed with non-null effect and non-null timeline Can't find variable: KeyframeEffectReadOnly
FAIL Animation can be constructed with non-null effect and no timeline parameter Can't find variable: KeyframeEffectReadOnly
FAIL Animation constructed by an effect with null target runs normally Can't find variable: KeyframeEffectReadOnly
FAIL Animation constructed with a keyframe that target element is in iframe Can't find variable: KeyframeEffectReadOnly
